feat: add directive-helpers and refactor
