+++ .gitignore
@@ -15,6 +15,8 @@
 /directive.js*
 /async-directive.d.ts*
 /async-directive.js*
+/directive-helpers.d.ts*
+/directive-helpers.js*
 /test/**/*.d.ts
 /test/**/*.d.ts.map
 /test/**/*.js
